requirements.txt
python-dotenv~=1.0.0
py-cord~=2.5.0
pymongo~=4.6.3